Reformatting of blocks | new aliases for version control and move into new directories | option-robust implemention of cmv

+++ shell/util/display.sh
@@ -2,7 +2,8 @@
 
 
 # Script only for Linux Systems
-if ! [[ "$OSTYPE" == "linux-gnu" ]]; then
+if ! [[ "$OSTYPE" == "linux-gnu" ]];
+then
   return 0
 fi
 
@@ -15,7 +16,8 @@ export PATH="$PATH:$HOME/System/sway"
 # GNOME setup
 function gnome() 
 {
-  if [ "$1" == "exit" ]; then
+  if [ "$1" == "exit" ];
+  then
     gnome-session-quit --logout --no-prompt
   else
     XDG_SESSION_TYPE=wayland dbus-run-session gnome-session
@@ -25,32 +27,39 @@ function gnome()
 # Switch between window-managers and TTY
 function mode()
 {
-  if [ -z "$1" ]; then
+  if [ -z "$1" ];
+  then
     echo "ERROR: Must provide a window manager to switch to"
     return 1
   fi
 
   # exit directly to tty
-  if [[ "$1" == "tty" ]]; then
-    for manager in "${managers[@}]}"; do
-      if pgrep -x "$1" > /dev/null; then
-        if ! $manager exit 2> /dev/null; then
+  if [[ "$1" == "tty" ]];
+  then
+    for manager in "${managers[@}]}";
+    do
+      if pgrep -x "$1" > /dev/null && ! $manager exit 2> /dev/null;
+      then
           pkill -x "$manager"
-        fi
       fi
     done
   fi
 
-  for manager in "${managers[@}]}"; do
-    if [[ "$1" == "$manager" ]]; then
+  for manager in "${managers[@}]}";
+  do
+    if [[ "$1" == "$manager" ]];
+    then
 
       # if already running check before relaunching
-      if pgrep -x "$1" > /dev/null; then
+      if pgrep -x "$1" > /dev/null;
+      then
         read -p "$manager is already running. Do you want to relaunch? [y/N]: " choice
 
         # affirmitve relaunch
-        if [[ "$choice" == "y" || "$choice" == "Y" ]]; then
-          if ! $manager exit 2> /dev/null; then
+        if [[ "$choice" == "y" || "$choice" == "Y" ]];
+        then
+          if ! $manager exit 2> /dev/null;
+          then
             pkill -x "$manager"
           fi
           exec "$@"
@@ -63,9 +72,12 @@ function mode()
 
       # switch to manager if not already running
       else   
-        for other in "${managers[@}]}"; do
-          if pgrep -x "$1" > /dev/null; then
-            if ! $other exit 2> /dev/null; then
+        for other in "${managers[@}]}";
+        do
+          if pgrep -x "$1" > /dev/null;
+          then
+            if ! $other exit 2> /dev/null;
+            then
               pkill -x "$other"
             fi
           fi
